add utility struct to include codes to check input values
